`test_non_python_task` moved to `test_full_workflow.py`
